rename tests directory
